requirements.txt
Django==3.1.1
django-admin-sortable2==0.7.7
django-tinymce-3.0.2
Pillow==7.2.0
